ref: 타입 변경, 메서드명 변경 등
리팩토링이라 쓰고 설계오류라 읽는다...

+++ src/main/java/kr/hhplus/be/server/application/user/UserPointFacade.java
@@ -14,14 +14,14 @@ public class UserPointFacade {
     private final UserPointService userPointService;
     private final PointHistoryService pointHistoryService;
 
-    public User chargePoint(Long userId, long amount) {
+    public User chargePoint(Long userId, int amount) {
 
         User updated = userPointService.chargePoint(userId, amount);  // chargePoint() 메서드 호출
         pointHistoryService.saveCharge(userId, amount);
         return updated;
     }
 
-    public User usePoint(Long userId, long amount) {
+    public User usePoint(Long userId, int amount) {
         // 포인트 사용 후 User 객체 반환
         User updated = userPointService.usePoint(userId, amount);
         // 포인트 내역 저장
@@ -29,7 +29,7 @@ public class UserPointFacade {
         return updated;
     }
 
-    public User refundPoint(Long userId, long amount, Long orderId) {
+    public User refundPoint(Long userId, int amount, Long orderId) {
         // 환불은 충전과 동일하게 처리
         User updated = userPointService.refundPoint(userId, amount);
         // 포인트 내역 저장
